TINKERPOP-2407 Added deserialization support for dict as keys
